Update Windows installation instructions
I just made a clean installation with a 32-bit Ruby on a Windows 8 system and these were the steps I needed to take in order to get it to work. The old instructions were a bit misleading as copying `aspell.dll` somewhere else in `PATH` leads to problems locating the installed dictionaries – the dicts are looked for in `<DLL path>/../dict` and `<DLL path>/../data` (both folders are needed).

I also capitalized a few names.